requirements.txt
appdirs==1.4.3
astroid==2.3.3
autoflake==1.3.1
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
colorama==0.4.3
cssselect==1.1.0
cycler==0.10.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==0.1.11
feedparser==5.2.1
google-api-core==1.17.0
google-auth==1.14.2
googleapis-common-protos==1.51.0
googlefinance==0.7
httplib2==0.17.3
idna==2.9
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.5.0
matplotlib==3.2.1
mccabe==0.6.1
multitasking==0.0.9
Naked==0.1.31
numpy==1.18.2
oauthlib==3.1.0
opencv-python==4.2.0.32
pandas==1.0.3
pandas-datareader==0.8.1
parse==1.15.0
Pillow==7.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyee==7.0.1
pyflakes==2.2.0
pylint==2.4.4
pyparsing==2.4.7
pyppeteer==0.0.25
pyquery==1.4.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==2.0
tqdm==4.45.0
uritemplate==3.0.1
urllib3==1.25.8
w3lib==1.21.0
websockets==8.1
wrapt==1.11.2
yahoo-fin==0.8.4
yfinance==0.1.54
